Prefer cache key from package.json
When a user provides only a package.json file we will generate a
package-lock.json file for them. This commit fixes a bug where the cache
key is retrieved both before and after generating this file; when
retrieved after, it has changed, causing false cache misses. Now we only
use the cache key from the package.json file in this case.

Added a test case from the sample app at
https://github.com/GoogleCloudPlatform/nodejs-docs-samples/tree/master/appengine/hello-world/standard